+++ ziti/cmd/ziti/cmd/edge_controller/create_edge_router.go
@@ -95,7 +95,7 @@ func runCreateEdgeRouter(o *createEdgeRouterOptions) error {
 }
 
 func getEdgeRouterJwt(o *createEdgeRouterOptions, id string) error {
-	list, err := listEntitiesOfType("edge-routers", nil, o.OutputJSONResponse, o.Out)
+	list, _, err := listEntitiesOfType("edge-routers", nil, o.OutputJSONResponse, o.Out)
 	if err != nil {
 		return err
 	}